+++ src/Contracts/Manager.php
@@ -52,16 +52,18 @@ interface Manager
     public function groups(array $groups): Manager;
 
     /**
-     * Get the count of passing checks.
+     * Filter checks by check code.
      *
-     * @since 0.1.0
+     * @since 0.5.0
      */
-    public function passing(): int;
+    public function code(string $code): Manager;
 
     /**
-     * Determine whether all checks are passing.
+     * Filter checks by a list of codes.
      *
-     * @since 0.1.0
+     * @since 0.5.0
+     *
+     * @param string[] $codes
      */
-    public function isPassing(): bool;
+    public function codes(array $codes): Manager;
 }