+++ tests/Unit/CheckResultTest.php
@@ -1,6 +1,7 @@
 <?php
 
 use Gerardojbaez\PhpCheckup\CheckResult;
+use Gerardojbaez\PhpCheckup\Status;
 use Gerardojbaez\PhpCheckup\Type;
 use PHPUnit\Framework\TestCase;
 
@@ -12,14 +13,15 @@ class CheckResultTest extends TestCase
         $type = Type::critical();
 
         $result = new CheckResult(
-            'My check', $type, true, 'Failing'
+            'My check', 'code', $type, Status::passing(), 'Failing'
         );
 
         // Assert
         $this->assertSame('My check', $result->name());
+        $this->assertSame('code', $result->code());
         $this->assertSame('Failing', $result->message());
         $this->assertSame($type, $result->type());
-        $this->assertTrue($result->isPassing());
-        $this->assertFalse($result->isFailing());
+        $this->assertTrue($result->status()->isPassing());
+        $this->assertFalse($result->status()->isFailing());
     }
 }